requirements.txt
fastapi
uvicorn[standard]
python-dotenv
SQLAlchemy
mysqlclient
aiomysql
PyMySQL
pydantic-settings
cryptography